python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptius-gui-schema"
version = "0.1.0"
description = "Type-safe UI schema definitions for cross-platform UI generation"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8.1"
authors = [
    {name = "Kanishk Gupta", email = "51890299+kanishkgupta2000@users.noreply.github.com"},
]
maintainers = [
    {name = "Kanishk Gupta", email = "51890299+kanishkgupta2000@users.noreply.github.com"},
]
keywords = ["ui", "schema", "typescript", "react", "vue", "angular", "pydantic", "validation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: User Interfaces",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/AgentBossMode/promptius-gui"
Repository = "https://github.com/AgentBossMode/promptius-gui"
Documentation = "https://github.com/AgentBossMode/promptius-gui#readme"
"Bug Tracker" = "https://github.com/AgentBossMode/promptius-gui/issues"

[project.optional-dependencies]
dev = [
    "langsmith>=0.1.147",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "datamodel-code-generator>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["promptius_gui_schema"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "fastapi>=0.120.0",
    "langchain-openai>=0.1.25",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.33.0",
]
